Add Automator service to copy permalink to Clipboard https://bugs.webkit.org/show_bug.cgi?id=170978
Reviewed by Joseph Pecoraro.

It is helpful to reference using a hyperlink a particular line of code when having
a discussion on IRC or in a bug. You can get such a link by navigating to the file
in the Trac Browse Source viewer and selecting the line your are interested in.
I found myself doing this often enough that I wrote an Automator service to do it
for me.

This Automator service works with Xcode 8 and Xcode 9 beta 5 (9M202q) or later.

* CopyPermalink/Copy WebKit Permalink.workflow/Contents/Info.plist: Added.
* CopyPermalink/Copy WebKit Permalink.workflow/Contents/document.wflow: Added.
* CopyPermalink/README: Added.

Identifier: 192228@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@220705 268f45cc-cd09-0410-ab3c-d52691b4dbfc