undefined URL in PopoverTracker for failed step https://bugs.webkit.org/show_bug.cgi?id=175593 <rdar://problem/33898769>
Reviewed by Daniel Bates.

This fix is a workaround for Buildbot 0.9 bug <https://github.com/buildbot/buildbot/issues/3529> where it
does not provide a URL to stdio for a build step. Buildbot 0.8 always included a URL to the stdio of the step.

* BuildSlaveSupport/build.webkit.org-config/public_html/dashboard/Scripts/BuildbotIteration.js:
(BuildbotIteration.prototype._parseData): Pass the URL while instantiating BuildbotTestResults.
(BuildbotIteration.prototype._urlForStep): Method to get the URL for a step.
* BuildSlaveSupport/build.webkit.org-config/public_html/dashboard/Scripts/BuildbotTestResults.js:
(BuildbotTestResults): Accept URL in constructor.


Identifier: 192275@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@220771 268f45cc-cd09-0410-ab3c-d52691b4dbfc