Regression(r221059): NetworkDataTask::didReceiveResponse() should not use PolicyUse for HTTP/0.9 https://bugs.webkit.org/show_bug.cgi?id=175872
Reviewed by Daniel Bates.

r221059 was calling the ResponseCompletionHandler with {} which ended up being
PolicyUse. Since this is an error case and do not want to receive the data, it
makes more sense to use PolicyIgnore instead. There should not be a behavior
change on WebKit side though since we are cancelling the load right after
calling the completion handler anyway.

Tests under http/tests/security/http-0.9/ are still passing.

* NetworkProcess/NetworkDataTask.cpp:
(WebKit::NetworkDataTask::didReceiveResponse):


Identifier: 192536@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@221081 268f45cc-cd09-0410-ab3c-d52691b4dbfc