[iOS] ViewServices started by StoreKitUIService may get suspended unexpectedly https://bugs.webkit.org/show_bug.cgi?id=175922 <rdar://problem/34028921>
Reviewed by Tim Horton.

ViewServices started by StoreKitUIService may get suspended unexpectedly. This is
because StoreKitUIService is itself a ViewService and is therefore in the background.
To address the issue, treat StoreKitUIService as initially foreground in
ApplicationStateTracker so that its ViewServices do not get suspended. We don't
have good way to deal with ViewServices started by other ViewServices unfortunately.

* Platform/spi/ios/UIKitSPI.h:
* UIProcess/ApplicationStateTracker.mm:
(WebKit::ApplicationStateTracker::ApplicationStateTracker):


Identifier: 192590@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@221138 268f45cc-cd09-0410-ab3c-d52691b4dbfc