REGRESSION (r220052): ASSERTION FAILED: !frame().isMainFrame() || !needsStyleRecalcOrLayout()  in WebCore::FrameView::updateLayoutAndStyleIfNeededRecursive() https://bugs.webkit.org/show_bug.cgi?id=175270
Reviewed by Simon Fraser.

Source/WebCore:

* dom/Document.cpp:
(WebCore::Document::Document): Initialize m_styleRecalcTimer with a lamdba so it can work
with a function that returns a bool and ignore the return value.
(WebCore::Document::updateStyleIfNeeded): Added a boolean return value indicating if the
function did any work or not.
* dom/Document.h: Updated for above change.

* page/FrameView.cpp:
(WebCore::appendRenderedChildren): Added helper that will later replace the
FrameView::renderedChildFrameViews function and is used below.
(WebCore::FrameView::updateLayoutAndStyleIfNeededRecursive): Instead of always doing two
passes of style and layout update do up to four passes, but stop as soon as a pass does
no work. This is slightly more efficient in cases where no layout and style update is
needed, and works correctly when a third pass is needed, which is what happens in the
test that was failing. We can eventually improve this further, but this resolves the
immediate problem we are seeing in the test.

LayoutTests:

* platform/mac-wk2/TestExpectations: Re-enable the disabled test.

Identifier: 192591@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@221139 268f45cc-cd09-0410-ab3c-d52691b4dbfc